Rename write.rs as artifact.rs, move Artifact struct to artifact.rs from read.rs
