+++ ocipkg/src/image/artifact.rs
@@ -2,7 +2,9 @@
 
 use crate::{
     digest::Digest,
-    image::{Artifact, Config, OciArchive, OciArchiveBuilder, OciArtifactBuilder},
+    image::{
+        Config, ImageLayout, OciArchive, OciArchiveBuilder, OciArtifact, OciArtifactBuilder, OciDir,
+    },
     media_types::{self, config_json},
     ImageName,
 };
@@ -11,6 +13,7 @@ use flate2::{write::GzEncoder, Compression};
 use std::{
     collections::HashMap,
     fs,
+    ops::{Deref, DerefMut},
     path::{Path, PathBuf},
 };
 
@@ -86,5 +89,42 @@ impl Builder {
             HashMap::new(),
         )?;
         self.builder.build()
+    }
+}
+
+/// ocipkg artifact defined as `application/vnd.ocipkg.v1.artifact`
+pub struct Artifact<Base: ImageLayout> {
+    base: OciArtifact<Base>,
+}
+
+impl<Base: ImageLayout> Deref for Artifact<Base> {
+    type Target = Base;
+    fn deref(&self) -> &Self::Target {
+        &self.base
+    }
+}
+
+impl<Base: ImageLayout> DerefMut for Artifact<Base> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.base
+    }
+}
+
+impl<Base: ImageLayout> Artifact<Base> {
+    pub fn new(base: Base) -> Self {
+        // TODO: Check media type
+        Self {
+            base: OciArtifact::new(base),
+        }
+    }
+
+    /// Get list of files stored in the ocipkg artifact
+    pub fn files(&mut self) -> Result<Vec<PathBuf>> {
+        todo!()
+    }
+
+    /// Unpack ocipkg artifact into local filesystem with `.oci-dir` directory
+    pub fn unpack(&mut self, dest: &Path) -> Result<OciDir> {
+        todo!()
     }
 }
